♻️[refactor] 회원 관련 엔티티 및 레포지토리 리팩토링
