+++ src/main/java/com/PickOne/domain/preference/controller/PreferenceController.java
@@ -17,10 +17,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
-@RestController()
+@RestController
 @RequestMapping("/api/preference")
 @RequiredArgsConstructor
-
 public class PreferenceController {
     private final PreferenceService preferenceService;
