+++ src/main/java/com/PickOne/domain/preference/dto/PreferenceRegisterDto.java
@@ -3,7 +3,7 @@ package com.PickOne.domain.preference.dto;
 import com.PickOne.domain.preference.model.entity.Genre;
 import com.PickOne.domain.preference.model.entity.Mbti;
 import com.PickOne.domain.preference.model.entity.Preference;
-import com.PickOne.domain.user.model.entity.Member;
+import com.PickOne.domain.user.model.Member;
 import java.util.List;
 import lombok.Builder;
 import lombok.Getter;
