+++ src/main/java/com/PickOne/domain/preference/model/UserInstrument.java
@@ -3,7 +3,7 @@ package com.PickOne.domain.preference.model;
 import com.PickOne.domain.preference.model.entity.Instrument;
 import com.PickOne.domain.preference.model.entity.Preference;
 import com.PickOne.domain.preference.model.entity.Proficiency;
-import com.PickOne.domain.user.model.entity.Member;
+import com.PickOne.domain.user.model.Member;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
