+++ src/main/java/com/PickOne/domain/preference/service/PreferenceService.java
@@ -10,7 +10,7 @@ import com.PickOne.domain.preference.model.entity.Preference;
 import com.PickOne.domain.preference.repository.PreferenceRepository;
 import com.PickOne.domain.preference.repository.UserGenreRepository;
 import com.PickOne.domain.preference.repository.UserInstrumentRepository;
-import com.PickOne.domain.user.model.entity.Member;
+import com.PickOne.domain.user.model.Member;
 import com.PickOne.domain.user.repository.MemberRepository;
 import com.PickOne.global.exception.BusinessException;
 import com.PickOne.global.exception.ErrorCode;
@@ -24,7 +24,6 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 @Slf4j
-
 public class PreferenceService {
     private final PreferenceRepository preferenceRepository;
     private final MemberRepository memberRepository;
@@ -34,10 +33,10 @@ public class PreferenceService {
 
     @Transactional
     public PreferenceResponseDto loadPreference(Long memberId) {
-        Member member = memberRepository.findByMemberId(memberId)     //시큐리티 적용후 memberId로 변경예정
+        Member member = memberRepository.findById(memberId)     //시큐리티 적용후 memberId로 변경예정
                 .orElseThrow(() -> new BusinessException(ErrorCode.USER_INFO_NOT_FOUND));
 
-        Preference preference = preferenceRepository.findByMemberMemberId(memberId)
+        Preference preference = preferenceRepository.findByMemberId(memberId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.PREFERENCE_NOT_FOUND));
 
         return PreferenceResponseDto.fromEntity(preference);
@@ -49,10 +48,10 @@ public class PreferenceService {
             PreferenceRegisterDto preferenceRegisterDto
             , Long memberId) {
 
-        Member member = memberRepository.findByMemberId(memberId)
+        Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new EntityNotFoundException("Member not found with ID: " + memberId));
 
-        preferenceRepository.findByMemberMemberId(memberId)
+        preferenceRepository.findByMemberId(memberId)
                 .ifPresent(preference -> {
                     throw new IllegalStateException("이미 존재합니다.");
                 });
@@ -66,10 +65,10 @@ public class PreferenceService {
             UserInstrumentRequestDto userInstrumentRequestDto
             , Long memberId) {
 
-        Member member = memberRepository.findByMemberId(memberId)
+        Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.USER_INFO_NOT_FOUND));
 
-        Preference preference = preferenceRepository.findByMemberMemberId(memberId)
+        Preference preference = preferenceRepository.findByMemberId(memberId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.PREFERENCE_NOT_FOUND));
 
         List<UserInstrument> userInstruments =userInstrumentRequestDto.toEntityList(member,preference);
@@ -81,10 +80,10 @@ public class PreferenceService {
             UserGenreRequestDto userGenreRequestDto
             , Long memberId) {
 
-        Member member = memberRepository.findByMemberId(memberId)
+        Member member = memberRepository.findById(memberId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.USER_INFO_NOT_FOUND));
 
-        Preference preference = preferenceRepository.findByMemberMemberId(memberId)
+        Preference preference = preferenceRepository.findByMemberId(memberId)
                 .orElseThrow(() -> new BusinessException(ErrorCode.PREFERENCE_NOT_FOUND));
 
         List<UserGenre> userGenres = userGenreRequestDto.toEntityList(member, preference);
